requirements.txt
psycopg2
python-telegram-bot
docxtpl
openai
